refactor: (#50) pages 디렉터리 페이지 단위 구조로 정리
